requirements.txt
awkward
inquirer
ipywidgets
matplotlib
numba
numba_stats
pandas
numpy
plotly
rich
scipy 
uproot
click
dash
dash_bootstrap_components
jupyter
kaleido
tqdm
xrootd
iminuit
pydantic
h5py
hickle
python-gssapi
paramiko
